Refactor out Checkpoint fixture
