+++ tests/integration/cloud/end_to_end/conftest.py
@@ -18,8 +18,13 @@ from great_expectations.execution_engine import SqlAlchemyExecutionEngine
 if TYPE_CHECKING:
     import py
 
-    from great_expectations.compatibility import pyspark
-    from great_expectations.compatibility.sqlalchemy import engine
+    from great_expectations.checkpoint import Checkpoint
+    from great_expectations.compatibility import pyspark, sqlalchemy
+    from great_expectations.core import ExpectationSuite
+    from great_expectations.datasource.fluent import (
+        BatchRequest,
+        DataAsset,
+    )
 
 LOGGER: Final = logging.getLogger("tests")
 
@@ -34,6 +39,52 @@ def context() -> CloudDataContext:
     )
     assert isinstance(context, CloudDataContext)
     return context
+
+
+@pytest.fixture(scope="module")
+def checkpoint(
+    context: CloudDataContext,
+    data_asset: DataAsset,
+    batch_request: BatchRequest,
+    expectation_suite: ExpectationSuite,
+    get_missing_checkpoint_error_type: type[Exception],
+) -> Iterator[Checkpoint]:
+    checkpoint_name = f"{data_asset.datasource.name} | {data_asset.name}"
+    _ = context.add_checkpoint(
+        name=checkpoint_name,
+        validations=[
+            {
+                "expectation_suite_name": expectation_suite.expectation_suite_name,
+                "batch_request": batch_request,
+            },
+            {
+                "expectation_suite_name": expectation_suite.expectation_suite_name,
+                "batch_request": batch_request,
+            },
+        ],
+    )
+    _ = context.add_or_update_checkpoint(
+        name=checkpoint_name,
+        validations=[
+            {
+                "expectation_suite_name": expectation_suite.expectation_suite_name,
+                "batch_request": batch_request,
+            }
+        ],
+    )
+    checkpoint = context.get_checkpoint(name=checkpoint_name)
+    assert (
+        len(checkpoint.validations) == 1
+    ), "Checkpoint was not updated in the previous method call."
+    yield checkpoint
+    # PP-691: this is a bug
+    # you should only have to pass name
+    context.delete_checkpoint(
+        # name=checkpoint_name,
+        id=checkpoint.ge_cloud_id,
+    )
+    with pytest.raises(get_missing_checkpoint_error_type):
+        context.get_checkpoint(name=checkpoint_name)
 
 
 @pytest.fixture(scope="module")
@@ -85,7 +136,7 @@ def table_factory() -> Iterator[TableFactory]:
     all_created_tables: dict[
         str, list[dict[Literal["table_name", "schema_name"], str | None]]
     ] = {}
-    engines: dict[str, engine.Engine] = {}
+    engines: dict[str, sqlalchemy.engine.Engine] = {}
 
     def _table_factory(
         gx_engine: SqlAlchemyExecutionEngine,
